add listener for the changes on the ExoPlayer to keep it in sync with the future Media Session.
Information from Udacity: Now we have a customized media player that works perfectly for our app. The last step is to make it play well with external clients and other apps. But before we can set up a Media Session, we need to find out what the state of ExoPlayer is so that we can keep our Media Session in sync.